+++ src/constants.js
@@ -1,4 +1,6 @@
 const aum_report_url =
   "https://www.mufap.com.pk/aum_report.php?tab=01&fname=&amc=&cat=&strdate=&endate=&submitted=&mnt=&yrs=&s=";
+const payout_url =
+  "https://www.mufap.com.pk/payout-report.php?tab=01&fname=&amc=&cat=&strdate=&endate=&submitted=&mnt=&yrs=&s=";
 
-module.exports = { aum_report_url };
+module.exports = { aum_report_url, payout_url };
