+++ src/server.js
@@ -1,44 +1,20 @@
 const puppeteer = require("puppeteer");
 const fs = require("fs/promises");
-const { scrapeFundTypes } = require("./scraper-functions/scrapeFundTypes");
-const { scrapeAMCs } = require("./scraper-functions/scrapeAMCs");
-const {
-  scrapeFundCategories,
-} = require("./scraper-functions/scrapeFundCategories");
-const { scrapeFunds } = require("./scraper-functions/funds/scrapeFunds");
 
 const { aum_report_url } = require("./constants");
+const { scrapeCommonEntities, scrapeFunds } = require("./scraper-functions");
 
 const startScraping = async () => {
   const browser = await puppeteer.launch();
   const page = await browser.newPage();
   await page.goto(aum_report_url);
-
-  const fundTypesTable = await scrapeFundTypes({ page });
-
-  const amcsTablePromise = scrapeAMCs({
+  const { fundTypesTable, amcsTable, fundCategoriesTable } =
+    await scrapeCommonEntities({ page, browser });
+  const fundsTable = await scrapeFunds({
     page,
     browser,
     linksToScrape: fundTypesTable?.map(({ link }) => link),
   });
-
-  const fundCategoriesTablePromise = scrapeFundCategories({
-    page,
-    browser,
-    linksToScrape: fundTypesTable?.map(({ link }) => link),
-  });
-
-  const fundsTablePromise = scrapeFunds({
-    page,
-    browser,
-    linksToScrape: fundTypesTable?.map(({ link }) => link),
-  });
-
-  const [amcsTable, fundCategoriesTable, fundsTable] = await Promise.all([
-    amcsTablePromise,
-    fundCategoriesTablePromise,
-    fundsTablePromise,
-  ]);
 
   await browser.close();
